Update stand. output/error file name to *.geos.log
The file for the combined standard output and standard error log has historically had the ending *geos.log. Update the default to include this for continuity (e.g. for automated scripts that look for a *geos.log) and to differentiate the file values other  *.log files that may be outputted (e.g. HEMCO.log).